requirements.txt
llama-index
llama_hub
chromadb
html2text
pinecone-client
transformers
nltk
